requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cloudpickle==2.0.0
cmake==3.22.1
cycler==0.11.0
Cython==0.29.26
dataclasses==0.6
datasets==1.17.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
entrypoints==0.4
filelock==3.4.1
flatbuffers==2.0
fonttools==4.29.1
frozenlist==1.2.0
fsspec==2022.1.0
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
huggingface-hub==0.4.0
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==6.9.1
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
jieba==0.42.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
llvmlite==0.38.0
loguru==0.6.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==5.2.0
multiprocess==0.70.12.2
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.6.7
notebook==6.4.8
numba==0.55.1
numpy==1.21.5
oauthlib==3.2.0
openprompt==0.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pkg_resources==0.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.3
ptyprocess==0.7.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.2
Pygments==2.11.2
pyparsing==3.0.6
pyrsistent==0.18.1
python-dateutil==2.8.2
pytorch-crf==0.7.0
pytorch-nlp==0.5.0
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
regex==2021.11.10
requests==2.27.1
requests-oauthlib==1.3.1
revtok==0.0.3
rouge==1.0.0
rsa==4.8
sacremoses==0.0.47
scikit-learn==0.20.2
scipy==1.4.1
seaborn==0.11.2
Send2Trash==1.8.0
sentencepiece==0.1.96
shap==0.40.0
six==1.16.0
sklearn==0.0
slicer==0.0.7
tbparse==0.0.6
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.9.0
torchmetrics==0.7.2
torchnlp==0.0.0.1
torchtext==0.4.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.16.2
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.5.2
wrapt==1.13.3
xxhash==2.0.2
yacs==0.1.8
yarl==1.7.2
zipp==3.6.0
